Split agent code, choose forage amount by regression
