+++ internal/clients/team1/client.go
@@ -3,12 +3,11 @@ package team1
 
 import (
 	"fmt"
-	"math"
-	"math/rand"
 
 	"github.com/SOMAS2020/SOMAS2020/internal/common/baseclient"
 	"github.com/SOMAS2020/SOMAS2020/internal/common/gamestate"
 	"github.com/SOMAS2020/SOMAS2020/internal/common/shared"
+	// "github.com/SOMAS2020/SOMAS2020/internal/clients/team1/foraging"
 )
 
 const id = shared.Team1
@@ -142,256 +141,6 @@ func (c *client) StartOfTurn() {
 	}
 	c.Logf("I'm all alone :c")
 
-}
-
-/********************/
-/***  Desperation   */
-/********************/
-
-func (c *client) desperateForage() shared.ForageDecision {
-	forageDecision := shared.ForageDecision{
-		Type:         bestROIForageType(c.forageHistory),
-		Contribution: c.gameState().ClientInfo.Resources,
-	}
-	c.Logf("[Forage][Decision]: Desperate | Decision %v", forageDecision)
-	return forageDecision
-}
-
-func (c *client) CommonPoolResourceRequest() shared.Resources {
-	switch c.emotionalState() {
-	case Anxious:
-		c.Logf("Common pool request: 20")
-		return 20
-	default:
-		return 0
-	}
-}
-
-func (c *client) RequestAllocation() shared.Resources {
-	var allocation shared.Resources
-
-	if c.emotionalState() == Desperate {
-		allocation = c.config.desperateStealAmount
-	} else if c.allocation != 0 {
-		allocation = c.allocation
-		c.allocation = 0
-	}
-
-	if allocation != 0 {
-		c.Logf("Taking %v from common pool", allocation)
-	}
-	return allocation
-}
-
-/********************/
-/***  Messages      */
-/********************/
-
-func (c *client) ReceiveCommunication(
-	sender shared.ClientID,
-	data map[shared.CommunicationFieldName]shared.CommunicationContent,
-) {
-	for field, content := range data {
-		switch field {
-		case shared.TaxAmount:
-			// if !__isPresident__(sender) {
-			// }
-			c.taxAmount = shared.Resources(content.IntegerData)
-		case shared.AllocationAmount:
-			// if !__isPresident__(sender) {
-			// }
-			c.allocation = shared.Resources(content.IntegerData)
-		}
-	}
-}
-
-func (c *client) GetTaxContribution() shared.Resources {
-	if c.config.evadeTaxes {
-		return 0
-	}
-	if c.gameState().Turn == 1 {
-		// Put in some initial resources to get IIGO to run
-		return c.config.kickstartTaxPercent * c.gameState().ClientInfo.Resources
-	}
-	return c.taxAmount
-}
-
-/********************/
-/***    Foraging    */
-/********************/
-
-func (c *client) randomForage() shared.ForageDecision {
-	// Up to 10% of our current resources
-	forageContribution := shared.Resources(0.1*rand.Float64()) * c.gameState().ClientInfo.Resources
-
-	var forageType shared.ForageType
-	if rand.Float64() < 0.5 {
-		forageType = shared.DeerForageType
-	} else {
-		forageType = shared.FishForageType
-	}
-
-	c.Logf("[Forage][Decision]:Random")
-	// TODO Add fish foraging when it's done
-	return shared.ForageDecision{
-		Type:         forageType,
-		Contribution: forageContribution,
-	}
-}
-
-// bestROIForageType finds the ForageType that has the best average ROI. If all
-// forageTypes have negative returns then it returns shared.ForageType(-1)
-func bestROIForageType(forageHistory ForageHistory) shared.ForageType {
-	bestForageType := shared.ForageType(-1)
-	bestROI := 0.0
-
-	for forageType, outcomes := range forageHistory {
-		ROIsum := 0.0
-		for _, outcome := range outcomes {
-			if outcome.contribution != 0 {
-				ROIsum += outcome.ROI()
-			}
-		}
-
-		averageROI := ROIsum / float64(len(outcomes))
-
-		if averageROI > bestROI {
-			bestROI = averageROI
-			bestForageType = forageType
-		}
-	}
-
-	return bestForageType
-}
-
-func bestROIOutcome(history []ForageOutcome) ForageOutcome {
-	bestOutcome := ForageOutcome{}
-
-	for _, outcome := range history {
-		if outcome.profit() >= 10 && outcome.contribution != 0 && outcome.ROI() > bestOutcome.ROI() {
-			bestOutcome = outcome
-		}
-	}
-	return bestOutcome
-}
-
-func (c *client) normalForage() shared.ForageDecision {
-	forageType := bestROIForageType(c.forageHistory)
-	if forageType == shared.ForageType(-1) {
-		// Here we have found that the best idea is to not forage
-		return shared.ForageDecision{
-			Type:         shared.FishForageType,
-			Contribution: 0,
-		}
-	}
-	expectedOutcome := bestROIOutcome(c.forageHistory[forageType])
-
-	contribution := expectedOutcome.contribution
-	// Cap to 20% of resources
-	contribution = shared.Resources(math.Min(
-		float64(contribution),
-		c.config.forageContributionCapPercent*float64(c.gameState().ClientInfo.Resources),
-	))
-	// Add some noise
-	contribution += shared.Resources(math.Min(
-		rand.Float64(),
-		c.config.forageContributionNoisePercent*float64(c.gameState().ClientInfo.Resources),
-	))
-
-	forageDecision := shared.ForageDecision{
-		Type:         forageType,
-		Contribution: contribution,
-	}
-	c.Logf(
-		"[Forage][Decision]: Decision %v | Expected ROI %v",
-		forageDecision, expectedOutcome.ROI())
-
-	return forageDecision
-}
-
-func (c *client) DecideForage() (shared.ForageDecision, error) {
-	if c.forageHistorySize() < c.config.randomForageTurns {
-		return c.randomForage(), nil
-	} else if c.emotionalState() == Desperate {
-		return c.desperateForage(), nil
-	} else {
-		return c.normalForage(), nil
-	}
-}
-
-func (c *client) ForageUpdate(forageDecision shared.ForageDecision, revenue shared.Resources) {
-	c.forageHistory[forageDecision.Type] = append(c.forageHistory[forageDecision.Type], ForageOutcome{
-		contribution: forageDecision.Contribution,
-		revenue:      revenue,
-		turn:         c.gameState().Turn,
-	})
-
-	c.Logf(
-		"[Forage][Update]: ForageType %v | Profit %v | Contribution %v | Actual ROI %v",
-		forageDecision.Type,
-		revenue-forageDecision.Contribution,
-		forageDecision.Contribution,
-		(revenue/forageDecision.Contribution)-1,
-	)
-}
-
-/********************/
-/***    IIFO        */
-/********************/
-
-func (c *client) MakeForageInfo() shared.ForageShareInfo {
-	var shareTo []shared.ClientID
-
-	for id, status := range c.gameState().ClientLifeStatuses {
-		if status != shared.Dead {
-			shareTo = append(shareTo, id)
-		}
-	}
-
-	lastDecisionTurn := -1
-	var lastDecision shared.ForageDecision
-	var lastRevenue shared.Resources
-
-	for forageType, outcomes := range c.forageHistory {
-		for _, outcome := range outcomes {
-			if int(outcome.turn) > lastDecisionTurn {
-				lastDecisionTurn = int(outcome.turn)
-				lastDecision = shared.ForageDecision{
-					Type:         forageType,
-					Contribution: outcome.contribution,
-				}
-				lastRevenue = outcome.revenue
-			}
-		}
-	}
-
-	if lastDecisionTurn < 0 {
-		shareTo = []shared.ClientID{}
-	}
-
-	forageInfo := shared.ForageShareInfo{
-		ShareTo:          shareTo,
-		ResourceObtained: lastRevenue,
-		DecisionMade:     lastDecision,
-	}
-
-	c.Logf("Sharing forage info: %v", forageInfo)
-	return forageInfo
-}
-
-func (c *client) ReceiveForageInfo(forageInfos []shared.ForageShareInfo) {
-	for _, forageInfo := range forageInfos {
-		c.forageHistory[forageInfo.DecisionMade.Type] =
-			append(
-				c.forageHistory[forageInfo.DecisionMade.Type],
-				ForageOutcome{
-					participant:  forageInfo.SharedFrom,
-					contribution: forageInfo.DecisionMade.Contribution,
-					revenue:      forageInfo.ResourceObtained,
-					turn:         c.gameState().Turn - 1,
-				},
-			)
-	}
 }
 
 /********************/
