+++ internal/clients/team1/client_test.go
@@ -1,6 +1,8 @@
 package team1
 
 import (
+//	"github.com/sajari/regression"
+	"math"
 	"testing"
 
 	"github.com/SOMAS2020/SOMAS2020/internal/common/gamestate"
@@ -23,34 +25,32 @@ func MakeTestClient(gamestate gamestate.ClientGameState) client {
 	return *c.(*client)
 }
 
-func TestRequestsResourcesWhenAnxious(t *testing.T) {
-	c := MakeTestClient(gamestate.ClientGameState{
-		ClientInfo: gamestate.ClientInfo{
-			Resources: 99,
-		},
-	})
-	c.config.anxietyThreshold = 100
+func TestRegression(t *testing.T) {
+	expectCoeffs := []float64{6.0, 2.0, 5.0}
 
-	gotRequest := c.CommonPoolResourceRequest()
-	if gotRequest <= 0 {
-		t.Errorf("Client did not make a resource request while anxious (%v)", gotRequest)
-	}
-}
-
-func TestStealsResourcesWhenDesperate(t *testing.T) {
-	c := MakeTestClient(gamestate.ClientGameState{
-		ClientInfo: gamestate.ClientInfo{
-			LifeStatus: shared.Critical,
-		},
-	})
-	c.config.desperateStealAmount = 50
-
-	gotTakeAmt := c.RequestAllocation()
-	if gotTakeAmt != c.config.desperateStealAmount {
-		t.Errorf(
-			`Client took the wrong amount while desperate:
-				Expected: %v
-				Got: %v`, c.config.desperateStealAmount, gotTakeAmt)
+	outcomes := []ForageOutcome{}
+	for i := 0.0; i < 50; i++ {
+		outcomes = append(outcomes, ForageOutcome{
+			contribution: shared.Resources(i),
+			revenue: shared.Resources(
+				expectCoeffs[2]*math.Pow(i, 2.0) +
+					expectCoeffs[1]*i +
+					expectCoeffs[0],
+			),
+		})
 	}
 
+	gotRegression := outcomeRegression(outcomes)
+
+	for i, expectedCoeff := range expectCoeffs {
+		gotCoeff := gotRegression.Coeff(i)
+		diff := math.Abs(expectedCoeff - gotCoeff)
+		if diff >= 0.0000001 {
+			t.Errorf("Incorrect coeffecient %v:\n\tExpected: %v\n\tGot: %v",
+				i,
+				expectedCoeff,
+				gotCoeff,
+			)
+		}
+	}
 }
