+++ internal/clients/team1/iifo.go
@@ -1,0 +1,60 @@
+package team1
+import "github.com/SOMAS2020/SOMAS2020/internal/common/shared"
+/********************/
+/***    IIFO        */
+/********************/
+
+func (c *client) MakeForageInfo() shared.ForageShareInfo {
+	var shareTo []shared.ClientID
+
+	for id, status := range c.gameState().ClientLifeStatuses {
+		if status != shared.Dead {
+			shareTo = append(shareTo, id)
+		}
+	}
+
+	lastDecisionTurn := -1
+	var lastDecision shared.ForageDecision
+	var lastRevenue shared.Resources
+
+	for forageType, outcomes := range c.forageHistory {
+		for _, outcome := range outcomes {
+			if int(outcome.turn) > lastDecisionTurn {
+				lastDecisionTurn = int(outcome.turn)
+				lastDecision = shared.ForageDecision{
+					Type:         forageType,
+					Contribution: outcome.contribution,
+				}
+				lastRevenue = outcome.revenue
+			}
+		}
+	}
+
+	if lastDecisionTurn < 0 {
+		shareTo = []shared.ClientID{}
+	}
+
+	forageInfo := shared.ForageShareInfo{
+		ShareTo:          shareTo,
+		ResourceObtained: lastRevenue,
+		DecisionMade:     lastDecision,
+	}
+
+	c.Logf("Sharing forage info: %v", forageInfo)
+	return forageInfo
+}
+
+func (c *client) ReceiveForageInfo(forageInfos []shared.ForageShareInfo) {
+	for _, forageInfo := range forageInfos {
+		c.forageHistory[forageInfo.DecisionMade.Type] =
+			append(
+				c.forageHistory[forageInfo.DecisionMade.Type],
+				ForageOutcome{
+					participant:  forageInfo.SharedFrom,
+					contribution: forageInfo.DecisionMade.Contribution,
+					revenue:      forageInfo.ResourceObtained,
+					turn:         c.gameState().Turn - 1,
+				},
+			)
+	}
+}
